pass through attachments in collection: https://github.com/metanorma/metanorma/issues/166
