+++ lib/metanorma/compile.rb
@@ -2,9 +2,10 @@ require "fileutils"
 require "nokogiri"
 require "htmlentities"
 require "yaml"
-
 require "fontist"
 require "fontist/manifest/install"
+require_relative "fontist"
+require_relative "compile_validate"
 
 module Metanorma
   class Compile
@@ -30,18 +31,14 @@ module Metanorma
     end
 
     def require_libraries(options)
-      if options[:require]
-        options[:require].each do |r|
-          require r
-        end
-      end
+      options&.dig(:require)&.each { |r| require r }
     end
 
     def xml_options_extract(file)
       xml = Nokogiri::XML(file)
       if xml.root
         @registry.root_tags.each do |k, v|
-          return { type: k }  if v == xml.root.name
+          return { type: k } if v == xml.root.name
         end
       end
       {}
@@ -63,48 +60,6 @@ module Metanorma
       options
     end
 
-    def validate_type(options)
-      unless options[:type]
-        Util.log("[metanorma] Error: Please specify a standard type: #{@registry.supported_backends}.", :error)
-        return nil
-      end
-
-      stdtype = options[:type].to_sym
-      metanorma_flavor = "metanorma-#{stdtype}"
-
-      unless @registry.supported_backends.include? stdtype
-        Util.log("[metanorma] Info: Loading `#{metanorma_flavor}` gem for standard type `#{stdtype}`.", :info)
-      end
-
-      begin
-        require "metanorma-#{stdtype}"
-        Util.log("[metanorma] Info: gem `#{metanorma_flavor}` loaded.", :info)
-
-      rescue Gem::ConflictError
-        Util.log("[metanorma] Error: Couldn't resolve dependencies for `metanorma-#{stdtype}`, Please add it to your Gemfile and run bundle install first", :error)
-        return false
-
-      rescue LoadError
-        Util.log("[metanorma] Error: loading gem `#{metanorma_flavor}` failed. Exiting.", :error)
-        return false
-      end
-
-      unless @registry.supported_backends.include? stdtype
-        Util.log("[metanorma] Error: The `#{metanorma_flavor}` gem still doesn't support `#{stdtype}`. Exiting.", :error)
-        return false
-      end
-
-      true
-    end
-
-    def validate_format(options)
-      unless options[:format] == :asciidoc
-        Util.log("[metanorma] Error: Only source file format currently supported is 'asciidoc'.", :error)
-        return false
-      end
-      true
-    end
-
     def get_extensions(options)
       options[:extension_keys] ||= @processor.output_formats.reduce([]) do |memo, (k, _)|
         memo << k
@@ -121,7 +76,7 @@ module Metanorma
       end
       if !extensions.include?(:presentation) and extensions.any? { |e| @processor.use_presentation_xml(e) }
         extensions << :presentation
-        end
+      end
       extensions
     end
 
@@ -187,7 +142,7 @@ module Metanorma
       xml.xpath("//sourcecode | //xmlns:sourcecode").each_with_index do |s, i|
         filename = s["filename"] || sprintf("sourcecode-%04d.txt", i)
         File.open("#{dirname}/sourcecode/#{filename}", "w:UTF-8") do |f|
-          f.write clean_sourcecode(s.dup) 
+          f.write clean_sourcecode(s.dup)
         end
       end
     end
@@ -197,6 +152,7 @@ module Metanorma
       FileUtils.mkdir_p "#{dirname}/image"
       xml.xpath("//image | //xmlns:image").each_with_index do |s, i|
         next unless /^data:image/.match s["src"]
+
         %r{^data:image/(?<imgtype>[^;]+);base64,(?<imgdata>.+)$} =~ s["src"]
         filename = s["filename"] || sprintf("image-%04d.%s", i, imgtype)
         File.open("#{dirname}/image/#{filename}", "wb") do |f|
@@ -222,22 +178,22 @@ module Metanorma
 
     # dependency ordering
     def sort_extensions_execution(ext)
-       case ext
-       when :xml then 0
-       when :rxl then 1
-       when :presentation then 2
-       else
-         99
-       end
+      case ext
+      when :xml then 0
+      when :rxl then 1
+      when :presentation then 2
+      else
+        99
+      end
     end
 
     def wrap_html(options, file_extension, outfilename)
-      if options[:wrapper] and /html$/.match file_extension
-          outfilename = outfilename.sub(/\.html$/, "")
-          FileUtils.mkdir_p outfilename
-          FileUtils.mv "#{outfilename}.html", outfilename
-          FileUtils.mv "#{outfilename}_images", outfilename, force: true
-        end
+      if options[:wrapper] && /html$/.match(file_extension)
+        outfilename = outfilename.sub(/\.html$/, "")
+        FileUtils.mkdir_p outfilename
+        FileUtils.mv "#{outfilename}.html", outfilename
+        FileUtils.mv "#{outfilename}_images", outfilename, force: true
+      end
     end
 
     # isodoc is Raw Metanorma XML
@@ -254,7 +210,8 @@ module Metanorma
         outfilename = f.sub(/\.[^.]+$/, ".#{file_extension}")
         if ext == :pdf
           font_locations = fontist_font_locations(options)
-          isodoc_options[:mn2pdf] = { font_manifest_file: font_locations.path } if font_locations
+          font_locations and
+            isodoc_options[:mn2pdf] = { font_manifest_file: font_locations.path }
         end
         if ext == :rxl
           options[:relaton] = outfilename
@@ -276,92 +233,7 @@ module Metanorma
       end
     end
 
-    def install_fonts(options)
-      if options[:no_install_fonts]
-        Util.log("[fontist] Skip font installation because" \
-          " --no-install-fonts argument passed", :debug)
-        return
-      end
-
-      if missing_fontist_manifest?
-        Util.log("[fontist] Skip font installation because font_manifest is missing", :debug)
-        return
-      end
-
-      @updated_formulas_repo = false
-
-      manifest = @processor.fonts_manifest
-      agree_to_terms = options[:agree_to_terms] || false
-      continue_without_fonts = options[:continue_without_fonts] || false
-      no_progress = options[:no_progress] || false
-
-      install_fonts_safe(manifest, agree_to_terms, continue_without_fonts, no_progress)
-    end
-
     private
-
-    def install_fonts_safe(manifest, agree, continue, no_progress)
-      fontist_install(manifest, agree, no_progress)
-    rescue Fontist::Errors::LicensingError
-      if continue
-        Util.log(
-          "[fontist] Processing will continue without fonts installed",
-          :debug
-        )
-      else
-        Util.log("[fontist] Aborting without proper fonts installed," \
-          " make sure that you have set option --agree-to-terms", :fatal)
-      end
-    rescue Fontist::Errors::FontError => e
-      log_level = continue ? :warning : :fatal
-      Util.log("[fontist] '#{e.font}' font is not supported. " \
-        "Please report this issue at github.com/metanorma/metanorma-"\
-        "#{@processor.short}/issues to report this issue.", log_level)
-    rescue Fontist::Errors::FormulaIndexNotFoundError
-      if @updated_formulas_repo
-        Util.log(
-          "[fontist] Bug: formula index not found after 'fontist update'",
-          :fatal
-        )
-      end
-      Util.log("[fontist] Missing formula index. Fetching it...", :debug)
-      Fontist::Formula.update_formulas_repo
-      @updated_formulas_repo = true
-      install_fonts_safe(manifest, agree, continue, no_progress)
-    end
-
-    def fontist_install(manifest, agree, no_progress)
-      Fontist::Manifest::Install.from_hash(
-        manifest,
-        confirmation: agree ? "yes" : "no",
-        no_progress: no_progress
-      )
-    end
-
-    def fontist_font_locations(options)
-      return nil if missing_fontist_manifest? || options[:no_install_fonts]
-
-      dump_fontist_manifest_locations(@processor.fonts_manifest)
-    rescue Fontist::Errors::FormulaIndexNotFoundError
-      raise unless options[:continue_without_fonts]
-
-      nil
-    end
-
-    def dump_fontist_manifest_locations(manifest)
-      location_manifest = Fontist::Manifest::Locations.from_hash(
-        manifest
-      )
-      location_manifest_file = Tempfile.new(["fontist_locations", ".yml"])
-      location_manifest_file.write location_manifest.to_yaml
-      location_manifest_file.flush
-
-      location_manifest_file
-    end
-
-    def missing_fontist_manifest?
-      !@processor.respond_to?(:fonts_manifest) || @processor.fonts_manifest.nil?
-    end
 
     # @param options [Hash]
     # @return [String]
