+++ spec/fixtures/collection/collection1.yml
@@ -42,6 +42,12 @@ manifest:
       docref:
         fileref: rice-amd.final.xml
         identifier: ISO 17301-1:2016/Amd.1:2017
+    - level: attachments
+      title: Attachments
+      docref:
+        - fileref: pics/action_schemaexpg1.svg
+          identifier: action_schemaexpg1.svg
+          attachment: true
 prefatory-content: 
 |
