+++ spec/input_spec.rb
@@ -2,25 +2,24 @@ require_relative "spec_helper"
 require "fileutils"
 
 RSpec.describe Metanorma::Input::Asciidoc do
-it "aborts if include error" do
-  begin
-    require "metanorma-iso"
-    expect{ Metanorma::Input::Asciidoc.new().process(<<~INPUT, "test.adoc", :iso) }.to raise_error(SystemExit)
-   = Document title
-  Author
-  :docfile: test.adoc
-  :nodoc:
+  it "aborts if include error" do
+    begin
+      require "metanorma-iso"
+      expect { Metanorma::Input::Asciidoc.new.process(<<~INPUT, "test.adoc", :iso) }.to raise_error(SystemExit)
+         = Document title
+        Author
+        :docfile: test.adoc
+        :nodoc:
 
-  include::common/common-sections/00-abstract.adoc[]
+        include::common/common-sections/00-abstract.adoc[]
 
-  INPUT
-  rescue SystemExit
+      INPUT
+    rescue SystemExit
+    end
   end
-end
-
 
   it "extracts Asciidoctor document attributes" do
-    expect(Hash[Metanorma::Input::Asciidoc.new().extract_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
+    expect(Hash[Metanorma::Input::Asciidoc.new.extract_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
       = Document title
       Author
       :script: a
@@ -48,42 +47,41 @@ end
       :break-up-urls-in-tables: v
       :htmlstylesheet-override: x
       :wordstylesheet-override: y
-      INPUT
+    INPUT
       {:bodyfont=>"b", :break_up_urls_in_tables=>"v", :datauriimage=>false, :doctoclevels=>"s", :header=>"m", :headerfont=>"c", :hierarchical_assets=>"t", :htmlcoverpage=>"g", :htmlintropage=>"h", :htmlstylesheet=>"f", :htmlstylesheet_override=>"x", :htmltoclevels=>"r", :i18nyaml=>"e", :olstyle=>"q", :script=>"a", :scripts=>"i", :scripts_pdf=>"j", :standardstylesheet=>"l", :titlefont=>"d", :ulstyle=>"p", :use_xinclude=>"u", :wordcoverpage=>"n", :wordintropage=>"o", :wordstylesheet=>"k", :wordstylesheet_override=>"y"}
     OUTPUT
   end
 
-    it "extracts Asciidoctor document attributes with default values" do
-    expect(Hash[Metanorma::Input::Asciidoc.new().extract_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
-    = Document title
-    Author
-    :hierarchical-assets: 
-    :use-xinclude:
-    :break-up-urls-in-tables:
-INPUT
-{:break_up_urls_in_tables=>"true", :datauriimage=>true, :hierarchical_assets=>"true", :use_xinclude=>"true"}
-OUTPUT
-    end
+  it "extracts Asciidoctor document attributes with default values" do
+    expect(Hash[Metanorma::Input::Asciidoc.new.extract_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
+      = Document title
+      Author
+      :hierarchical-assets:#{' '}
+      :use-xinclude:
+      :break-up-urls-in-tables:
+    INPUT
+      {:break_up_urls_in_tables=>"true", :datauriimage=>true, :hierarchical_assets=>"true", :use_xinclude=>"true"}
+    OUTPUT
+  end
 
   it "extracts Asciidoctor document attributes for Metanorma" do
-    expect(Hash[Metanorma::Input::Asciidoc.new().extract_metanorma_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
+    expect(Hash[Metanorma::Input::Asciidoc.new.extract_metanorma_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
       = Document title
       Author
       :mn-document-class: a
       :mn-output-extensions: b
       :mn-keep-asciimath:
-      INPUT
+    INPUT
       {:asciimath=>true, :extensions=>"b", :type=>"a"}
     OUTPUT
   end
 
   it "extracts Asciidoctor document attributes for Metanorma" do
-    expect(Hash[Metanorma::Input::Asciidoc.new().extract_metanorma_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
+    expect(Hash[Metanorma::Input::Asciidoc.new.extract_metanorma_options(<<~"INPUT").sort].to_s + "\n").to eq <<~"OUTPUT"
       = Document title
       Author
-      INPUT
+    INPUT
       {}
     OUTPUT
   end
-
 end
