+++ spec/output_spec.rb
@@ -2,11 +2,11 @@ require_relative "spec_helper"
 require "fileutils"
 
 RSpec.describe Metanorma::Output::XslfoPdf do
-  ASSETS_DIR = 'spec/assets/outputtest'
-  INPUT = "#{ASSETS_DIR}/a.xml"
-  OUTPUT = "#{ASSETS_DIR}/a.pdf"
-  DROP = [OUTPUT]
-  XSL = "#{ASSETS_DIR}/iso.international-standard.xsl"
+  ASSETS_DIR = "spec/assets/outputtest".freeze
+  INPUT = "#{ASSETS_DIR}/a.xml".freeze
+  OUTPUT = "#{ASSETS_DIR}/a.pdf".freeze
+  DROP = [OUTPUT].freeze
+  XSL = "#{ASSETS_DIR}/iso.international-standard.xsl".freeze
 
   it "generates a PDF" do
     FileUtils.rm_f DROP
@@ -24,25 +24,25 @@ RSpec.describe Metanorma::Output::XslfoPdf do
 
     generator = Metanorma::Output::XslfoPdf.new
 
-    generator.convert(File.join(Dir.pwd, INPUT), File.join(Dir.pwd, OUTPUT), XSL)
+    generator.convert(File.join(Dir.pwd, INPUT), File.join(Dir.pwd, OUTPUT),
+                      XSL)
     expect(File.exist?(File.join(Dir.pwd, OUTPUT))).to be true
   end
 
-=begin
-  it "handle url as input" do
-    FileUtils.rm_f DROP
-    FileUtils.mkdir_p ASSETS_DIR
-
-    generator = Metanorma::Output::XslfoPdf.new
-
-    generator.convert("file://#{File.join(Dir.pwd, INPUT)}", OUTPUT, XSL)
-    expect(File.exist?(OUTPUT)).to be true
-  end
-=end
+  #   it "handle url as input" do
+  #     FileUtils.rm_f DROP
+  #     FileUtils.mkdir_p ASSETS_DIR
+  #
+  #     generator = Metanorma::Output::XslfoPdf.new
+  #
+  #     generator.convert("file://#{File.join(Dir.pwd, INPUT)}", OUTPUT, XSL)
+  #     expect(File.exist?(OUTPUT)).to be true
+  #   end
 
   it "Missing input path" do
     generator = Metanorma::Output::XslfoPdf.new
-    expect { generator.convert("random.html", "random.pdf", XSL) }.to raise_error
+    expect { generator.convert("random.html", "random.pdf", XSL) }
+      .to raise_error
   end
 
   it "Path with spaces" do
@@ -58,4 +58,3 @@ RSpec.describe Metanorma::Output::XslfoPdf do
     expect(File.exist?(output)).to be true
   end
 end
-
