+++ spec/registry_spec.rb
@@ -4,13 +4,13 @@ RSpec.describe Metanorma::Registry do
   it "has a registry" do
     class NewProcessor < Metanorma::Processor
       def initialize
+        super
         @short = :new
       end
 
-      def output_formats 
+      def output_formats
         { xyz: "xyz" }
       end
-
     end
     registry = Metanorma::Registry.instance
     registry.register(NewProcessor)
@@ -20,16 +20,16 @@ RSpec.describe Metanorma::Registry do
     expect(registry.output_formats[:new]&.to_s).to eq '{:xyz=>"xyz"}'
   end
 
-    it "deals with multiple aliases" do
+  it "deals with multiple aliases" do
     class NewProcessor < Metanorma::Processor
       def initialize
-        @short = [:new, :new2]
+        super
+        @short = %i[new new2]
       end
 
       def output_formats
         { xyz: "xyz" }
       end
-
     end
     registry = Metanorma::Registry.instance
     registry.register(NewProcessor)
@@ -43,15 +43,15 @@ RSpec.describe Metanorma::Registry do
     expect(registry.output_formats[:new2]&.to_s).to eq '{:xyz=>"xyz"}'
   end
 
-
   it "warns when registered class is not a Metanorma processor" do
     registry = Metanorma::Registry.instance
-    expect{registry.register(Metanorma)}.to raise_error(Error) 
+    expect { registry.register(Metanorma) }.to raise_error(Error)
   end
 
   it "detects root tag of iso" do
     class NewProcessor < Metanorma::Processor
       def initialize
+        super
         @short = :iso
         @asciidoctor_backend = :iso
       end
@@ -59,7 +59,6 @@ RSpec.describe Metanorma::Registry do
       def output_formats
         { xyz: "xyz" }
       end
-
     end
     require "metanorma-iso"
     registry = Metanorma::Registry.instance
